move SecretKey into its own module
Move-only, except for a couple import paths.
